+++ packages/common/authFunctions/package.json
@@ -4,7 +4,7 @@
   "description": "Common resources for auth functions",
   "author": "NHS Digital",
   "license": "MIT",
-  "main": "src/index.ts",
+  "main": "lib/src/index.js",
   "scripts": {
     "unit": "POWERTOOLS_DEV=true NODE_OPTIONS=--experimental-vm-modules jest --no-cache --coverage",
     "lint": "eslint  --max-warnings 0 --fix --config ../../../eslint.config.mjs .",
